refactor(`LaunchScreen`): Change the folder to view
